Revert "Completely new cross-platform API introduced"
This reverts commit 0def20c4bef2b242db15b38202a54735101f5791.

Temporarily reverting to fix long-tail issues with internal usage
patterns.
