+++ pkgs/platform/.gitignore
@@ -4,7 +4,8 @@
 .dart_tool/
 .pub/
 build/
-doc/
+packages
+.packages
 
 # Include when developing application packages.
 pubspec.lock
