--- pkgs/platform/lib/src/testing/test_platforms.dart
@@ -1,846 +1,0 @@
-// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-/// User-configurable platform libraries.
-///
-/// Should only be used for testing.
-// ignore: unnecessary_library_name - Used by DartDoc
-library test_platforms;
-
-import 'dart:collection';
-import 'dart:convert';
-
-import 'package:meta/meta.dart';
-
-import '../platforms_impl.dart';
-import '../util/json_keys.dart' as json_key;
-import 'zone_overrides.dart' as overrides;
-
-/// Custom Dart runtime platform information for testing.
-///
-/// Implements [Platform], but allows a [TestBrowserPlatform] or
-/// [TestNativePlatform] to be the non-`null` platform object.
-@visibleForTesting
-final class TestPlatform extends PlatformTestBase {
-  /// The current native platform, if running on a native platform.
-  @override
-  final TestNativePlatform? nativePlatform;
-
-  /// The current browser platform, if running on a browser platform.
-  @override
-  final TestBrowserPlatform? browserPlatform;
-
-  /// Whether currently running on a native platform.
-  @override
-  bool get isNative => nativePlatform != null;
-
-  /// Whether currently running in a browser.
-  @override
-  bool get isBrowser => browserPlatform != null;
-
-  /// A custom platform with the given [nativePlatform].
-  ///
-  /// Used by [TestNativePlatform.run] to create a [TestPlatform]
-  /// with itself as the [Platform.nativePlatform].
-  const TestPlatform._native(TestNativePlatform this.nativePlatform)
-    : browserPlatform = null;
-
-  /// A custom platform with the given [browserPlatform].
-  ///
-  /// Used by [TestBrowserPlatform.run] to create a [TestPlatform]
-  /// with itself as the [Platform.browserPlatform].
-  const TestPlatform._browser(TestBrowserPlatform this.browserPlatform)
-    : nativePlatform = null;
-
-  /// A platform with *neither* the native or the browser platform available.
-  @visibleForTesting
-  const TestPlatform.unknown() : nativePlatform = null, browserPlatform = null;
-
-  /// Creates a [TestPlatform] with the same properties as [platform].
-  @visibleForTesting
-  factory TestPlatform.fromPlatform(Platform platform) {
-    // There is currently no way to have a platform with both
-    // a browser platform and a native platform.
-    var native = platform.nativePlatform;
-    if (native != null) return TestPlatform.fromNative(native);
-    var browser = platform.browserPlatform;
-    if (browser != null) return TestPlatform.fromBrowser(browser);
-    return const TestPlatform.unknown();
-  }
-
-  /// Creates a [TestPlatform] with a new [TestNativePlatform].
-  ///
-  /// Creates a [TestNativePlatform.new] with the provided arguments,
-  /// and a [TestPlatform] with that as [nativePlatform].
-  ///
-  /// Equivalent to:
-  /// ```dart
-  /// TestPlatform.fromNative(TestNativePlatform(...))
-  /// ```
-  @visibleForTesting
-  TestPlatform.native({
-    Map<String, String>? environment,
-    String? executable,
-    List<String>? executableArguments,
-    String? lineTerminator,
-    String? localeName,
-    String? localHostname,
-    int? numberOfProcessors,
-    String? operatingSystem,
-    String? operatingSystemVersion,
-    String? packageConfig,
-    String? pathSeparator,
-    String? resolvedExecutable,
-    Uri? script,
-    bool? stdinSupportsAnsi,
-    bool? stdoutSupportsAnsi,
-    String? version,
-  }) : this._native(
-         TestNativePlatform(
-           environment: environment,
-           executable: executable,
-           executableArguments: executableArguments,
-           lineTerminator: lineTerminator,
-           localeName: localeName,
-           localHostname: localHostname,
-           numberOfProcessors: numberOfProcessors,
-           operatingSystem: operatingSystem,
-           operatingSystemVersion: operatingSystemVersion,
-           packageConfig: packageConfig,
-           pathSeparator: pathSeparator,
-           resolvedExecutable: resolvedExecutable,
-           script: script,
-           stdinSupportsAnsi: stdinSupportsAnsi,
-           stdoutSupportsAnsi: stdoutSupportsAnsi,
-           version: version,
-         ),
-       );
-
-  /// Creates a [TestPlatform] with a copy of a given [NativePlatform].
-  ///
-  /// Creates a [TestNativePlatform] from [nativePlatform],
-  /// as by [TestNativePlatform.from], and a [TestPlatform]
-  /// with that as its [Platform.nativePlatform].
-  @visibleForTesting
-  TestPlatform.fromNative(NativePlatform nativePlatform)
-    : this._native(TestNativePlatform.from(nativePlatform));
-
-  /// Creates a [TestPlatform] with a [TestNativePlatform] created from JSON.
-  ///
-  /// A new [TestNativePlatform] is created using [TestNativePlatform.fromJson]
-  /// with [nativePlatformJson] as argument, and a new [TestPlatform] is created
-  /// with that native platform as [Platform.nativePlatform].
-  ///
-  /// Equivalent to:
-  /// ```dart
-  /// TestPlatform.fromNative(TestNativePlatform.fromJson(nativePlatformJson))
-  /// ```
-  @visibleForTesting
-  TestPlatform.nativeFromJson(String nativePlatformJson)
-    : this._native(TestNativePlatform.fromJson(nativePlatformJson));
-
-  /// Creates a [TestPlatform] with a copy of a given [BrowserPlatform].
-  ///
-  /// Creates a [TestBrowserPlatform] from [browserPlatform],
-  /// as by [TestBrowserPlatform.from], and a [TestPlatform]
-  /// with that as its [Platform.browserPlatform].
-  @visibleForTesting
-  TestPlatform.fromBrowser(BrowserPlatform browserPlatform)
-    : this._browser(TestBrowserPlatform.from(browserPlatform));
-
-  /// Creates a [TestPlatform] with a [TestBrowserPlatform] created from JSON.
-  ///
-  /// A new [TestBrowserPlatform] is created using
-  /// [TestBrowserPlatform.fromJson] with [browserPlatformJson] as argument,
-  /// and a new [TestPlatform] is created with that native platform as
-  /// [Platform.browserPlatform].
-  @visibleForTesting
-  TestPlatform.browserFromJson(String browserPlatformJson)
-    : this._browser(TestBrowserPlatform.fromJson(browserPlatformJson));
-
-  /// Runs [testCode] with this platform as the current platform.
-  ///
-  /// While [testCode] is running, the [Platform.current] refers to this
-  /// platform object.
-  ///
-  /// Prior reads of [Platform.current] will retain their original value,
-  /// so the [testCode] should make sure to read [Platform.current]
-  /// when it's needed, and not rely on previously read values.
-  @visibleForTesting
-  R run<R>(R Function() testCode) =>
-      overrides.runWith(testCode, this, _OverrideMarker.marker);
-}
-
-/// Instance used to mark overrides as used.
-enum _OverrideMarker implements overrides.OverrideMarker { marker }
-
-/// A custom [BrowserPlatform] for testing.
-///
-/// The individual properties can be left unset by the constructors.
-/// Accessing such properties will throw, but a test which doesn't use
-/// those properties doesn't have to provide values for them.
-@visibleForTesting
-final class TestBrowserPlatform extends BrowserPlatformTestBase {
-  static const _className = 'TestBrowserPlatform';
-
-  final String? _version;
-  final String? _userAgent;
-
-  /// Creates a custom `TestBrowserPlatform` with the given properties.
-  ///
-  /// If an argument is omitted or `null`, the created object will throw
-  /// an error when reading the corresponding property.
-  @visibleForTesting
-  TestBrowserPlatform({String? version, String? userAgent})
-    : _version = version,
-      _userAgent = userAgent;
-
-  /// Creates a new [TestBrowserPlatform] with properties from a JSON string.
-  ///
-  /// The [jsonText] must be a valid JSON string representing a JSON object,
-  /// and the values for the [BrowserPlatform] properties are extracted from
-  /// the object's values for keys with the properties' names.
-  ///
-  /// Example:
-  /// ```dart
-  /// var testBrowserPlatform = TestBrowserPlatform.fromJson('''{
-  ///   "userAgent": "Banana/1.0 (BananaOS 0.9.1)",
-  /// }''')'
-  /// ```
-  /// The JSON object's value for such a key must be either a string or `null`.
-  ///
-  /// Throws a [FormatException] if the [jsonText] string is not valid
-  /// JSON text, the JSON text does not represent a JSON object,
-  /// or a property's value is not a string.
-  /// (The property values must be strings because all properties of
-  /// `TestBrowserPlatform` are strings).
-  ///
-  /// Other keys are ignored. Missing property keys,
-  /// or keys with a `null` value, leave the property undefined.
-  ///
-  /// The [toJson] method will leave out undefined properties, so
-  /// `TestBrowserPlatform.fromJson(TestBrowserPlatform.toJson())` can create
-  /// an exact copy at a later time.
-  /// (To create a copy _right now_, `TestBrowserPlatform.copyWith()`
-  /// is easier and more efficient.)
-  @visibleForTesting
-  factory TestBrowserPlatform.fromJson(String jsonText) {
-    var json = jsonDecode(jsonText);
-    if (json is! Map<String, Object?>) {
-      throw FormatException('Not a JSON map', jsonText);
-    }
-    return TestBrowserPlatform(
-      version: _getJsonProperty<String>(json, json_key.version, jsonText),
-      userAgent: _getJsonProperty<String>(json, json_key.userAgent, jsonText),
-    );
-  }
-
-  /// Creates a custom `TestBrowserPlatform` as a copy of [browserPlatform].
-  ///
-  /// If [browserPlatform] is a [TestBrowserPlatform], the new object is
-  /// an exact copy of the original, including any undefined properties.
-  /// Otherwise the new object is a [TestBrowserPlatform]
-  /// that has the same property values as [browserPlatform].
-  @visibleForTesting
-  factory TestBrowserPlatform.from(BrowserPlatform browserPlatform) {
-    if (browserPlatform is TestBrowserPlatform) {
-      return browserPlatform.copyWith();
-    }
-    return TestBrowserPlatform(
-      version: browserPlatform.version,
-      userAgent: browserPlatform.userAgent,
-    );
-  }
-
-  /// Creates a new [TestBrowserPlatform] from this one.
-  ///
-  /// If a parameter is given non-`null` argument value,
-  /// the created object will have that value for the corresponding property,
-  /// otherwise it will use this object's value for that property.
-  @visibleForTesting
-  TestBrowserPlatform copyWith({String? version, String? userAgent}) =>
-      TestBrowserPlatform(
-        version: version ?? _version,
-        userAgent: userAgent ?? _userAgent,
-      );
-
-  /// Runs [testCode] with this object as `Platform.current.browserPlatform`.
-  ///
-  /// While [testCode] is running, the [Platform.browserPlatform]
-  /// of [Platform.current] refers to this [TestBrowserPlatform].
-  ///
-  /// Prior reads of [Platform.current] will retain their original value,
-  /// so the `testCode` should make sure to read [Platform.current]
-  /// when it's needed, and not rely on previously read values.
-  @visibleForTesting
-  R run<R>(R Function() testCode) => overrides.runWith(
-    testCode,
-    TestPlatform._browser(this),
-    _OverrideMarker.marker,
-  );
-
-  /// A JSON-encoded representation of this browser platform test-configuration.
-  ///
-  /// Unset values are not included.
-  ///
-  /// Can be parsed back by [TestBrowserPlatform.fromJson].
-  @override
-  String toJson() => const JsonEncoder.withIndent(
-    '  ',
-  ).convert({json_key.version: ?_version, json_key.userAgent: ?_userAgent});
-
-  @override
-  String get userAgent =>
-      _throwIfUnset(_userAgent, _className, json_key.userAgent);
-
-  @override
-  String get version => _throwIfUnset(_version, _className, json_key.version);
-}
-
-/// A custom [NativePlatform] for testing.
-///
-/// The individual properties can be left unset by the constructors.
-/// Accessing such properties will throw, but a test which doesn't use
-/// those properties doesn't have to provide values for them.
-///
-/// This class should not be used in production, since it reduces the
-/// compiler's ability to predict the values of the
-/// [NativePlatform.operatingSystem] or, for example, [NativePlatform.isLinux]
-/// checks, which can lead to larger compiled programs.
-@visibleForTesting
-final class TestNativePlatform extends NativePlatformTestBase {
-  static const String _className = 'TestNativePlatform';
-
-  /// Operating system ID string, or `null` if not configured.
-  ///
-  /// Use one of the [NativePlatform.android] through [NativePlatform.windows]
-  /// constants.
-  final String? _operatingSystem;
-  final String? _operatingSystemVersion;
-  final Map<String, String>? _environment;
-  final String? _executable;
-  final List<String>? _executableArguments;
-  final String? _lineTerminator;
-  final String? _localeName;
-  final String? _localHostname;
-  final int? _numberOfProcessors;
-  final String? _pathSeparator;
-  @override
-  final String? packageConfig;
-  final String? _resolvedExecutable;
-  final Uri? _script;
-  final bool? _stdinSupportsAnsi;
-  final bool? _stdoutSupportsAnsi;
-  final String? _version;
-
-  /// Creates a new [TestNativePlatform] with the specified properties.
-  ///
-  /// Parameters that are not provided with a non-`null` value
-  /// will leave the corresponding property without a value.
-  ///
-  /// Reading, for example, [TestNativePlatform.operatingSystem]
-  /// when it had not been given a value, will throw an error.
-  /// This behavior can be used in tests to ensure that code does not
-  /// read properties that it is not supposed to.
-  ///
-  /// The [operatingSystem] string must be one of the known operating system
-  /// ID strings in [NativePlatform.operatingSystemValues].
-  @visibleForTesting
-  TestNativePlatform({
-    Map<String, String>? environment,
-    String? executable,
-    List<String>? executableArguments,
-    String? lineTerminator,
-    String? localeName,
-    String? localHostname,
-    int? numberOfProcessors,
-    String? operatingSystem,
-    String? operatingSystemVersion,
-    this.packageConfig,
-    String? pathSeparator,
-    String? resolvedExecutable,
-    Uri? script,
-    bool? stdinSupportsAnsi,
-    bool? stdoutSupportsAnsi,
-    String? version,
-  }) : _environment = environment == null
-           ? null
-           : operatingSystem == NativePlatform.windows
-           ? (_caseInsensitiveMap()..addAll(environment))
-           : environment,
-       _executable = executable,
-       _executableArguments = executableArguments,
-       _lineTerminator = lineTerminator,
-       _localeName = localeName,
-       _localHostname = localHostname,
-       _numberOfProcessors = numberOfProcessors,
-       _operatingSystem = operatingSystem,
-       _operatingSystemVersion = operatingSystemVersion,
-       _pathSeparator = pathSeparator,
-       _resolvedExecutable = resolvedExecutable,
-       _script = script,
-       _stdinSupportsAnsi = stdinSupportsAnsi,
-       _stdoutSupportsAnsi = stdoutSupportsAnsi,
-       _version = version;
-
-  /// Creates a new [TestNativePlatform] with properties from a JSON string.
-  ///
-  /// The [jsonText] must be a valid JSON string representing a JSON object,
-  /// and the values for the [NativePlatform] properties are extracted from
-  /// the object's values for keys with the properties' names.
-  ///
-  /// Example:
-  /// ```dart
-  /// var testNativePlatform = TestNativePlatform.fromJson('''{
-  ///   "operatingSystem": "linux",
-  ///   "operatingSystemVersion": "MockVersion.2.3",
-  ///   "numberOfProcessors": 42,
-  ///   "lineTerminator": "\r",
-  ///   "packageConfig": null
-  /// }''')'
-  /// ```
-  /// The JSON object's value for such a key must be either a string or `null`.
-  ///
-  /// Throws a [FormatException] if the [jsonText] string is not valid
-  /// JSON text, or the JSON text does not represent a JSON object.
-  /// The property values must have the correct type for the corresponding
-  /// [NativePlatform] property, or be `null`.
-  /// If the map value of a property key is not a string, integer, boolean,
-  /// list of strings, or map with string keys and values,
-  /// as required by the corresponding property,
-  /// a [FormatException] is also thrown.
-  ///
-  /// Other keys are ignored. Missing property keys,
-  /// or keys with a `null` value, leave the property undefined.
-  ///
-  /// The [toJson] method will leave out undefined properties, so
-  /// `TestNativePlatform.fromJson(TestNativePlatform.toJson())` can create
-  /// an exact copy at a later time.
-  /// (To create a copy _right now_, `TestNativePlatform.copyWith()`
-  /// is easier and more efficient.)
-  @visibleForTesting
-  factory TestNativePlatform.fromJson(String jsonText) {
-    Object? jsonObject = jsonDecode(jsonText);
-    if (jsonObject is! Map<String, Object?>) {
-      throw FormatException('Not a JSON object', jsonText);
-    }
-
-    // Read out the values that require special validation first.
-    var environment = _getJsonProperty<Map<String, Object?>>(
-      jsonObject,
-      json_key.environment,
-      jsonText,
-    );
-    var executableArguments = _getJsonProperty<List<Object?>>(
-      jsonObject,
-      json_key.executableArguments,
-      jsonText,
-    );
-    var script = _getJsonProperty<String>(jsonObject, 'script', jsonText);
-
-    return TestNativePlatform(
-      // Check that environment values are strings. Filter out null values.
-      environment: environment == null
-          ? null
-          : <String, String>{
-              for (var MapEntry(:key, :value) in environment.entries)
-                if (value != null)
-                  key: (value is String
-                      ? value
-                      : _failJsonParse<String>(
-                          'environment[$key]',
-                          value,
-                          jsonText,
-                        )),
-            },
-      executable: _getJsonProperty<String>(
-        jsonObject,
-        json_key.executable,
-        jsonText,
-      ),
-      // Check the executable arguments are strings.
-      executableArguments: executableArguments == null
-          ? null
-          : <String>[
-              for (var argument in executableArguments)
-                argument is String
-                    ? argument
-                    : _failJsonParse<String>(
-                        'executableArguments',
-                        argument,
-                        jsonText,
-                      ),
-            ],
-      lineTerminator: _getJsonProperty<String>(
-        jsonObject,
-        json_key.lineTerminator,
-        jsonText,
-      ),
-      localeName: _getJsonProperty<String>(
-        jsonObject,
-        json_key.localeName,
-        jsonText,
-      ),
-      localHostname: _getJsonProperty<String>(
-        jsonObject,
-        json_key.localHostname,
-        jsonText,
-      ),
-      numberOfProcessors: _getJsonProperty<int>(
-        jsonObject,
-        json_key.numberOfProcessors,
-        jsonText,
-      ),
-      operatingSystem: _getJsonProperty<String>(
-        jsonObject,
-        json_key.operatingSystem,
-        jsonText,
-      ),
-      operatingSystemVersion: _getJsonProperty<String>(
-        jsonObject,
-        json_key.operatingSystemVersion,
-        jsonText,
-      ),
-      packageConfig: _getJsonProperty<String>(
-        jsonObject,
-        json_key.packageConfig,
-        jsonText,
-      ),
-      pathSeparator: _getJsonProperty<String>(
-        jsonObject,
-        json_key.pathSeparator,
-        jsonText,
-      ),
-      resolvedExecutable: _getJsonProperty<String>(
-        jsonObject,
-        json_key.resolvedExecutable,
-        jsonText,
-      ),
-      // Checks that the script is a valid URI.
-      script: script == null ? null : Uri.parse(script),
-      stdinSupportsAnsi: _getJsonProperty<bool>(
-        jsonObject,
-        json_key.stdinSupportsAnsi,
-        jsonText,
-      ),
-      stdoutSupportsAnsi: _getJsonProperty<bool>(
-        jsonObject,
-        json_key.stdoutSupportsAnsi,
-        jsonText,
-      ),
-      version: _getJsonProperty<String>(jsonObject, json_key.version, jsonText),
-    );
-  }
-
-  /// Creates a new [TestNativePlatform] with the properties of [platform].
-  @visibleForTesting
-  factory TestNativePlatform.from(NativePlatform platform) =>
-      platform
-          is TestNativePlatform // Values may be unset.
-      ? platform.copyWith()
-      : TestNativePlatform(
-          environment: <String, String>{...platform.environment},
-          executable: platform.executable,
-          executableArguments: List.of(platform.executableArguments),
-          lineTerminator: platform.lineTerminator,
-          localeName: platform.localeName,
-          localHostname: platform.localHostname,
-          numberOfProcessors: platform.numberOfProcessors,
-          operatingSystem: platform.operatingSystem,
-          operatingSystemVersion: platform.operatingSystemVersion,
-          packageConfig: platform.packageConfig,
-          pathSeparator: platform.pathSeparator,
-          resolvedExecutable: platform.resolvedExecutable,
-          script: platform.script,
-          stdinSupportsAnsi: platform.stdinSupportsAnsi,
-          stdoutSupportsAnsi: platform.stdoutSupportsAnsi,
-          version: platform.version,
-        );
-
-  /// Whether the operating system is Android.
-  @override
-  bool get isAndroid => operatingSystem == NativePlatform.android;
-
-  /// Whether the operating system is Fuchsia.
-  @override
-  bool get isFuchsia => operatingSystem == NativePlatform.fuchsia;
-
-  /// Whether the operating system is iOS.
-  @override
-  bool get isIOS => operatingSystem == NativePlatform.iOS;
-
-  /// Whether the operating system is Linux.
-  @override
-  bool get isLinux => operatingSystem == NativePlatform.linux;
-
-  /// Whether the operating system is OS X.
-  @override
-  bool get isMacOS => operatingSystem == NativePlatform.macOS;
-
-  /// Whether the operating system is Windows.
-  @override
-  bool get isWindows => operatingSystem == NativePlatform.windows;
-
-  @override
-  String get operatingSystem =>
-      _throwIfUnset(_operatingSystem, _className, json_key.operatingSystem);
-
-  @override
-  String get operatingSystemVersion => _throwIfUnset(
-    _operatingSystemVersion,
-    _className,
-    json_key.operatingSystemVersion,
-  );
-
-  @override
-  Map<String, String> get environment =>
-      _throwIfUnset(_environment, _className, json_key.environment);
-
-  @override
-  String get executable =>
-      _throwIfUnset(_executable, _className, json_key.executable);
-
-  @override
-  List<String> get executableArguments => _throwIfUnset<List<String>>(
-    _executableArguments,
-    _className,
-    json_key.executableArguments,
-  );
-
-  @override
-  String get lineTerminator =>
-      _throwIfUnset(_lineTerminator, _className, json_key.lineTerminator);
-
-  @override
-  String get localeName =>
-      _throwIfUnset(_localeName, _className, json_key.localeName);
-
-  @override
-  String get localHostname =>
-      _throwIfUnset(_localHostname, _className, json_key.localHostname);
-
-  @override
-  int get numberOfProcessors => _throwIfUnset(
-    _numberOfProcessors,
-    _className,
-    json_key.numberOfProcessors,
-  );
-
-  @override
-  String get pathSeparator =>
-      _throwIfUnset(_pathSeparator, _className, json_key.pathSeparator);
-
-  @override
-  String get resolvedExecutable => _throwIfUnset(
-    _resolvedExecutable,
-    _className,
-    json_key.resolvedExecutable,
-  );
-
-  @override
-  Uri get script => _throwIfUnset(_script, _className, json_key.script);
-
-  @override
-  bool get stdinSupportsAnsi =>
-      _throwIfUnset(_stdinSupportsAnsi, _className, json_key.stdinSupportsAnsi);
-
-  @override
-  bool get stdoutSupportsAnsi => _throwIfUnset(
-    _stdoutSupportsAnsi,
-    _className,
-    json_key.stdoutSupportsAnsi,
-  );
-
-  @override
-  String get version => _throwIfUnset(_version, _className, json_key.version);
-
-  /// Creates a new [TestNativePlatform] from this one.
-  ///
-  /// If a parameter is given non-`null` argument value,
-  /// the created object will have that value for the corresponding property,
-  /// otherwise it will use this object's value for that property.
-  @visibleForTesting
-  TestNativePlatform copyWith({
-    Map<String, String>? environment,
-    String? executable,
-    List<String>? executableArguments,
-    String? lineTerminator,
-    String? localeName,
-    String? localHostname,
-    int? numberOfProcessors,
-    String? operatingSystem,
-    String? operatingSystemVersion,
-    String? packageConfig,
-    String? pathSeparator,
-    String? resolvedExecutable,
-    Uri? script,
-    bool? stdinSupportsAnsi,
-    bool? stdoutSupportsAnsi,
-    String? version,
-  }) {
-    return TestNativePlatform(
-      environment:
-          environment ??
-          (_environment == null ? null : <String, String>{..._environment}),
-      executable: executable ?? _executable,
-      executableArguments: (executableArguments ?? _executableArguments)
-          ?.toList(),
-      lineTerminator: lineTerminator ?? _lineTerminator,
-      localeName: localeName ?? _localeName,
-      localHostname: localHostname ?? _localHostname,
-      numberOfProcessors: numberOfProcessors ?? _numberOfProcessors,
-      operatingSystem: operatingSystem ?? _operatingSystem,
-      operatingSystemVersion: operatingSystemVersion ?? _operatingSystemVersion,
-      packageConfig: packageConfig ?? this.packageConfig,
-      pathSeparator: pathSeparator ?? _pathSeparator,
-      resolvedExecutable: resolvedExecutable ?? _resolvedExecutable,
-      script: script ?? _script,
-      stdinSupportsAnsi: stdinSupportsAnsi ?? _stdinSupportsAnsi,
-      stdoutSupportsAnsi: stdoutSupportsAnsi ?? _stdoutSupportsAnsi,
-      version: version ?? _version,
-    );
-  }
-
-  /// Migration helper for legacy `FakePlatform.copyWith`.
-  ///
-  /// Use [TestNativePlatform.copyWith] instead.
-  /// Work directly with [TestNativePlatform], rather than creating a
-  /// `TestPlatform` from it.
-  ///
-  /// Only works if there is a current [Platform.nativePlatform].
-  ///
-  /// Will be deprecated and removed when legacy classes are removed.
-  @visibleForTesting
-  TestPlatform copyPlatformWith({
-    int? numberOfProcessors,
-    String? pathSeparator,
-    String? operatingSystem,
-    String? operatingSystemVersion,
-    String? localHostname,
-    Map<String, String>? environment,
-    String? executable,
-    String? resolvedExecutable,
-    Uri? script,
-    List<String>? executableArguments,
-    String? packageConfig,
-    String? version,
-    bool? stdinSupportsAnsi,
-    bool? stdoutSupportsAnsi,
-    String? localeName,
-  }) {
-    return TestPlatform._native(
-      copyWith(
-        numberOfProcessors: numberOfProcessors,
-        pathSeparator: pathSeparator,
-        operatingSystem: operatingSystem,
-        operatingSystemVersion: operatingSystemVersion,
-        localHostname: localHostname,
-        environment: environment,
-        executable: executable,
-        resolvedExecutable: resolvedExecutable,
-        script: script,
-        executableArguments: executableArguments,
-        packageConfig: packageConfig,
-        version: version,
-        stdinSupportsAnsi: stdinSupportsAnsi,
-        stdoutSupportsAnsi: stdoutSupportsAnsi,
-        localeName: localeName,
-      ),
-    );
-  }
-
-  /// Runs [testCode] with this native platform as current native platform.
-  ///
-  /// While [testCode] is running, the [Platform.current] refers a new
-  /// [TestPlatform] with this native platform as [Platform.nativePlatform].
-  ///
-  /// Prior reads of [Platform.current] will retain their original value,
-  /// so the [testCode] should make sure to read [Platform.current]
-  /// when it's needed, and not rely on previously read values.
-  @visibleForTesting
-  R run<R>(R Function() testCode) => overrides.runWith(
-    testCode,
-    TestPlatform._native(this),
-    _OverrideMarker.marker,
-  );
-
-  /// Wraps [withOverride] to run it with this native platform as current.
-  ///
-  /// The returned function runs [withOverride] through [run] every time it's
-  /// called, to ensure all the calls will see the same override.
-  @visibleForTesting
-  R Function() bind<R>(R Function() withOverride) =>
-      () => run(withOverride);
-
-  /// A JSON-encoded representation of this native platform test-configuration.
-  ///
-  /// Unset values are not included.
-  ///
-  /// Can be parsed back by [TestNativePlatform.fromJson].
-  @override
-  String toJson() {
-    return const JsonEncoder.withIndent('  ').convert(<String, dynamic>{
-      json_key.environment: ?_environment,
-      json_key.executable: ?_executable,
-      json_key.executableArguments: ?_executableArguments,
-      json_key.lineTerminator: ?_lineTerminator,
-      json_key.localeName: ?_localeName,
-      json_key.localHostname: ?_localHostname,
-      json_key.numberOfProcessors: ?_numberOfProcessors,
-      json_key.operatingSystem: ?_operatingSystem,
-      json_key.operatingSystemVersion: ?_operatingSystemVersion,
-      json_key.packageConfig: ?packageConfig,
-      json_key.pathSeparator: ?_pathSeparator,
-      json_key.resolvedExecutable: ?_resolvedExecutable,
-      json_key.script: ?_script?.toString(),
-      json_key.stdinSupportsAnsi: ?_stdinSupportsAnsi,
-      json_key.stdoutSupportsAnsi: ?_stdoutSupportsAnsi,
-      json_key.version: ?_version,
-    });
-  }
-}
-
-/// Reports a failure to parse a JSON property at its expected type.
-Never _failJsonParse<T>(String property, Object? value, String? source) {
-  throw FormatException('Property "$property" is not a $T: $value', source);
-}
-
-/// Reads a property from a JSON map, and throws if it has the wrong type.
-///
-/// Reads the [property] key from the [jsonMap] and returns the value,
-/// if it has type `T?`.
-///
-/// A `null` value is allowed and treated the same
-/// as if the [property] key is absent in the [jsonMap].
-/// Only fails if the key exists and has a value of a different type.
-T? _getJsonProperty<T extends Object>(
-  Map<String, Object?> jsonMap,
-  String property,
-  String? source,
-) {
-  var value = jsonMap[property];
-  if (value is T?) return value;
-  _failJsonParse<T>(property, value, source);
-}
-
-/// Throws if [value] is `null`, otherwise returns it.
-///
-/// Used by getters on test-classes like [TestNativePlatform],
-/// which allow otherwise non-nullable properties to be unset
-/// as long as they are not read.
-/// This function is called when reading such a property.
-T _throwIfUnset<T extends Object>(T? value, String className, String name) =>
-    value ?? (throw StateError('$className.$name property is unset'));
-
-/// A simple inefficient case-insensitive map.
-///
-/// Used by [TestNativePlatform.environment] when the (user-configured)
-/// [NativePlatform.operatingSystem] is [NativePlatform.windows].
-/// Performance is not a consideration since [TestNativePlatform]
-/// is only intended for testing, and the environment is not expected
-/// to be larger than necessary.
-Map<String, String> _caseInsensitiveMap() => HashMap<String, String>(
-  equals: (String a, String b) => a == b || a.toLowerCase() == b.toLowerCase(),
-  hashCode: (String a) => a.toLowerCase().hashCode,
-);
