--- pkgs/platform/lib/src/util/json_keys.dart
@@ -1,26 +1,0 @@
-// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-/// The names of properties and the keys used by `toJson` and `fromJson`.
-/// As constants to avoid typos. In an import namespace to avoid pollution,
-/// and so that they can be used in tests.
-library;
-
-const String environment = 'environment';
-const String executable = 'executable';
-const String executableArguments = 'executableArguments';
-const String lineTerminator = 'lineTerminator';
-const String localeName = 'localeName';
-const String localHostname = 'localHostname';
-const String numberOfProcessors = 'numberOfProcessors';
-const String operatingSystem = 'operatingSystem';
-const String operatingSystemVersion = 'operatingSystemVersion';
-const String packageConfig = 'packageConfig';
-const String pathSeparator = 'pathSeparator';
-const String resolvedExecutable = 'resolvedExecutable';
-const String script = 'script';
-const String stdinSupportsAnsi = 'stdinSupportsAnsi';
-const String stdoutSupportsAnsi = 'stdoutSupportsAnsi';
-const String userAgent = 'userAgent';
-const String version = 'version';
