+++ pkgs/platform/pubspec.yaml
@@ -2,19 +2,16 @@ name: platform
 description: A pluggable, mockable platform information abstraction for Dart.
 repository: https://github.com/dart-lang/core/tree/main/pkgs/platform
 issue_tracker: https://github.com/dart-lang/core/issues
-version: 3.2.0-wip
+version: 3.1.7-wip
 
-environment:
-  # IF UPDATED: Also update SDK version in pubspecs in `example/` and `out/`.
-  sdk: ^3.10.0
-
-dependencies:
-  meta: ^1.18.0
-
-dev_dependencies:
-  dart_flutter_team_lints: ^3.5.2
-  path: ^1.9.1
-  test: ^1.29.0
 topics:
   - information
   - platform
+
+environment:
+  sdk: ^3.2.0
+
+dev_dependencies:
+  dart_flutter_team_lints: ^3.1.0
+  test: ^1.16.8
+
