--- pkgs/platform/test/browser_platform_test.dart
@@ -1,60 +1,0 @@
-// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-// Test of [browserPlatform].
-@OnPlatform({'!browser': Skip('Browser-only functionality')})
-library;
-
-import 'dart:js_interop'; // For `appVersion` and `navigator`.
-
-import 'package:platform/platform.dart';
-import 'package:test/test.dart';
-
-void main() {
-  const isNative = bool.fromEnvironment('dart.library.io');
-  const isBrowser = bool.fromEnvironment('dart.library.js_interop');
-  final platform = Platform.current;
-  final browserPlatform = platform.browserPlatform;
-
-  test('Running on browser', () {
-    expect(isBrowser, true, reason: 'dart.library.js_interop');
-    expect(isNative, false, reason: 'dart.library.io');
-    expect(platform.isNative, false, reason: 'Platform.current.isNative');
-    expect(platform.isBrowser, true, reason: 'Platform.current.isBrowser');
-    expect(browserPlatform, isNotNull, reason: 'BrowserPlatform.current');
-    expect(platform.nativePlatform, isNull, reason: 'NativePlatform.current');
-  });
-
-  if (browserPlatform == null) return; // Promote to non-null.
-
-  test('static properties', () {
-    // The `current` getter is the same objects as
-    // `BrowserPlatform.current`.
-    expect(
-      browserPlatform,
-      same(BrowserPlatform.current),
-      reason: 'BrowserPlatform.current',
-    );
-  });
-
-  test('Has same values as JS', () {
-    // Everything is taken from browser (`window.navigator`).
-    expect(
-      browserPlatform.userAgent,
-      userAgent,
-      reason: 'window.navigator.userAgent',
-    );
-    expect(
-      browserPlatform.version,
-      appVersion,
-      reason: 'window.navigator.appVersion',
-    );
-  });
-}
-
-@JS('window.navigator.appVersion')
-external String get appVersion;
-
-@JS('window.navigator.userAgent')
-external String get userAgent;
